NewMars2040Tradespace + Many Changes
I re-wrote the main function and the logistics module, as well as
confirming that the Surface stuff works with the Architecture and
Results.
Many Changes, run NewMars2040Tradespace.m now as the master.

Still need more de-bugging.
Migrated the Surface stuff to GitHub, assists with debug, and allows
others to help us.